+++ playwright/utils/mock-api.ts
@@ -9,3 +9,17 @@ export async function mockUsernameApi(page: Page, message: string) {
     });
   });
 }
+
+export async function mockCreatePatronApi(
+  page: Page,
+  name: string,
+  barcode: string
+) {
+  await page.route("**/library-card/api/create-patron", async (route) => {
+    await route.fulfill({
+      status: 200,
+      contentType: "application/json",
+      body: JSON.stringify({ name, barcode }),
+    });
+  });
+}